+++ trunk/TP1/TPAlgoritmos2D/EstrategiasDibujo/DDA.cs
@@ -11,26 +11,26 @@ namespace SistemasGraficos.EstrategiasDibujo
     {
         public static void DibujarLinea(Linea linea)
         {
-            float xInicio = linea.Inicio.GetXFlotante();
-            float xFin = linea.Fin.GetXFlotante();
-            float yInicio = linea.Inicio.GetYFlotante();
-            float yFin = linea.Fin.GetYFlotante();
+            double xInicio = linea.Inicio.GetXFlotante();
+            double xFin = linea.Fin.GetXFlotante();
+            double yInicio = linea.Inicio.GetYFlotante();
+            double yFin = linea.Fin.GetYFlotante();
 
-            float pendiente = linea.GetPendiente();
+            double pendiente = linea.GetPendiente();
 
-            float pasoY = 1;
-            float pasoX = 1;
+            double pasoY = 1;
+            double pasoX = 1;
 
             if (pendiente == 0) pasoY = 0;
-            else if (float.IsInfinity(pendiente)) pasoX = 0;
+            else if (double.IsInfinity(pendiente)) pasoX = 0;
             else if (Math.Abs(yFin - yInicio) > Math.Abs(xFin - xInicio)) pasoX = Math.Abs(1/pendiente);
             else pasoY = Math.Abs(pendiente);
 
             if (yFin < yInicio) pasoY = -pasoY;
             if (xFin < xInicio) pasoX = -pasoX;
 
-            float xIteracion = xInicio;
-            float yIteracion = yInicio;
+            double xIteracion = xInicio;
+            double yIteracion = yInicio;
 
             while (((xIteracion >= xInicio && xIteracion <= xFin) ||
                 (xIteracion <= xInicio && xIteracion >= xFin)) &&
